Do not hard code file with job descriptions
It's a somewhat bigger diff, but the actual code change is minor.  Most
stuff was just moved around to get the jobs in scope.

The big where clause is not necessarily beautiful, but it's better than
having the job file hard coded + we can always try to clean it up (e.g.
by moving stuff into a reader.